LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/page-load/svg-script-xlinkhref.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""


FAIL Speculative parsing, page load: svg-script-xlinkhref Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
